premiere version stable avec les principales fonctionalites

+++ src/Edahira/DahiraBundle/Controller/CaissesController.php
@@ -8,29 +8,21 @@ use Edahira\DahiraBundle\Form\CaissesType;
 
 class CaissesController extends Controller
 {
-    public function indexAction($page)
+    public function indexAction()
     {
-		if($page < 1){
-			throw $this->createNotFoundException('Page inexistente (page = '.$page.')');
-			$page = 1;
-		}
-
-		$caisses = $this->getDoctrine()
-						->getManager()
-						->getRepository('EdahiraDahiraBundle:Caisses')
-						->getCaisses(10, $page);
-						
+		$caisses = $this->get('security.context')->getToken()->getUser()->getActiveDahira()->getCaisses();		
+		
 		return $this->render('EdahiraDahiraBundle:Caisses:index.html.twig', array(
-		                     'caisses'   => $caisses,
-		                     'page'       => $page,
-		                     'nombrePage' => ceil(count($caisses)/10)
+		                     'caisses'   => $caisses
 		));
     }
 	
     public function editerAction(Caisses $caisse = null)
     {
+    	$user = $this->get('security.context')->getToken()->getUser();
+
 		if(is_null($caisse)){
-			$caisse = new Caisses();
+			$caisse = new Caisses($user->getActiveDahira());
 		}
 
 		$form = $this->createForm(new CaissesType(), $caisse);
@@ -38,37 +30,54 @@ class CaissesController extends Controller
 		$request = $this->get('request');
 
 		if($request->getMethod() == 'POST'){
+
 			$form->bind($request);
 				
 			if($form->isValid()){
+
+				$caisses = $user->getActiveDahira()->getCaisses();
+
+                foreach ($caisses as $caisseVerif){
+
+					if($caisse->getNom() == $caisseVerif->getNom() and $caisse->getId() != $caisseVerif->getId()){
+
+						$this->get('session')->getFlashBag()->add('warning','warning.caisse.exist');
+
+						return $this->render('EdahiraDahiraBundle:Caisses:editer.html.twig',
+							array('form'=>$form->createView(), 'id'=> $caisse->getId()
+							));
+					}
+				}
+
 				$em = $this->getDoctrine()
 				           ->getManager();
 				 
 				$em->persist($caisse);
 				$em->flush();
 
-				$this->get('session')->getFlashBag()->add('info','Caisse bien enregistré');
+				$this->get('session')->getFlashBag()->add('success','success.caisse.edeted');
 				return $this->redirect($this->generateUrl('caisse_index'));
 			}
 		}
 		/* 
 			*/
 		return $this->render('EdahiraDahiraBundle:Caisses:editer.html.twig',
-			array('form'=>$form->createView(), 'id'=> $caisse->getId()));
-    }
-
-    public function afficherAction()
-    {
+			array('form'=>$form->createView(), 'caisse'=> $caisse
+				));
     }
 
     public function supprimerAction(Caisses $caisse)
     {
-    	$em = $this->getDoctrine()->getManager();
-		$em->remove($caisse);
-		$em->flush();
-		$this->get('session')->getFlashBag()->add('info', 'Caisse supprimée avec succès');
-		
-		return $this->redirect($this->generateUrl('caisse_index'));
+    	if($this->get('request')->getMethod() == 'POST'){
+    		// var_dump($caisse);exit;
+	    	$em = $this->getDoctrine()->getManager();
+			$em->remove($caisse);
+			$em->flush();
+			$this->get('session')->getFlashBag()->add('success', 'action.caisse.deleted');
+			
+			return $this->redirect($this->generateUrl('caisse_index'));
+		}
+    	return $this->render('EdahiraDahiraBundle:Caisses:supprimer.html.twig', array('caisse' => $caisse));
     }
 
     public function changeretatAction(Caisses $caisse, $etat)
@@ -81,9 +90,9 @@ class CaissesController extends Controller
 		$em->persist($caisse);
 		$em->flush();
 		if($etat)
-		    $this->get('session')->getFlashBag()->add('info', 'Caisse fermé avec succès');
+	    	$this->get('session')->getFlashBag()->add('success', 'flash.caisse.opened');
 	    else
-	    	$this->get('session')->getFlashBag()->add('info', 'Caisse ouvert avec succès');
+		    $this->get('session')->getFlashBag()->add('success', 'flash.caisse.closed');
 		return $this->redirect( $this->generateUrl('caisse_index'));
     }
 }
